Create upload.php
Upload.php is a  self-submitting php page which uploads multiple image files to the server. It is best suited to be used inside a pop-up but can easily be modified as per any requirement. Simply add this file along with the css and an empty folder named "uploads" to get it working.